Rename to sparktoolbox, Writers, and Fetchers
