+++ src/main/scala/sparktoolbox/SparkPlatformTrait.scala
@@ -1,4 +1,4 @@
-package sparkutils
+package sparktoolbox
 
 import org.apache.spark.sql.SparkSession
