+++ src/test/scala/sparktoolbox/TestSparkPlatform.scala
@@ -1,4 +1,4 @@
-package sparkutils
+package sparktoolbox
 
 import org.apache.spark.sql.SparkSession
 import java.util.concurrent.ConcurrentHashMap
